# swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "API Library",
    "description": "Contain all api from kulony backend"
  },
  "servers": [
    {
      "url": "https://kulony-backend.herokuapp.com/"
    }
  ],
  "paths": {
...
    "/api/post/create": {
      "post": {
        "tags": [
          "Post"
        ],
        "description": "ส่ง post ที่ต้องการเพิ่มใน Database",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "topic_id": {
                    "example": "any"
                  },
                  "post_title": {
                    "example": "any"
                  },
                  "post_content": {
                    "example": "any"
                  },
                  "cover_photo_url": {
                    "example": "any"
                  },
                  "post_photo_url": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/post/all_post": {
      "get": {
        "tags": [
          "Post"
        ],
        "description": "ขอข้อมูล post ทั้งหมดจาก Database",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/post/{post_id}": {
      "get": {
        "tags": [
          "Post"
        ],
        "description": "ขอข้อมูลของ post_id นั้นๆ",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/post/{entity_id}/report": {
      "post": {
        "tags": [
          "Post"
        ],
        "description": "ส่ง Report post",
        "parameters": [
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "report_type": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/post/{post_id}/edit": {
      "put": {
        "tags": [
          "Post"
        ],
        "description": "แก้ไข post",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/post/like/{post_id}": {
      "post": {
        "tags": [
          "Post"
        ],
        "description": "Like post และส่ง notice ให้เจ้าของ Post",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/post/unlike/{post_id}": {
      "delete": {
        "tags": [
          "Post"
        ],
        "description": "Unlike post และลบ Notice ออก",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/reply/create": {
      "post": {
        "tags": [
          "Reply"
        ],
        "description": "ส่งข้อมูลสำหรับสร้าง Reply",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "comment_id": {
                    "example": "any"
                  },
                  "reply_content": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/reply/{comment_id}": {
      "get": {
        "tags": [
          "Reply"
        ],
        "description": "ขอ reply ทั้งหมดของ Comment นั้น",
        "parameters": [
          {
            "name": "comment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/reply/{entity_id}/report": {
      "post": {
        "tags": [
          "Reply"
        ],
        "description": "ส่ง report Reply",
        "parameters": [
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "report_type": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/reply/{reply_id}/edit": {
      "put": {
        "tags": [
          "Reply"
        ],
        "description": "แก้ไข Reply",
        "parameters": [
          {
            "name": "reply_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/reply/like/{reply_id}": {
      "put": {
        "tags": [
          "Reply"
        ],
        "description": "like reply",
        "parameters": [
          {
            "name": "reply_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/reply/unlike/{reply_id}": {
      "put": {
        "tags": [
          "Reply"
        ],
        "description": "unlike reply",
        "parameters": [
          {
            "name": "reply_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}